refactor(import): rename from-url to import-from-url
"from-url" is too vague standing alone in a file switcher; the explicit
"import-" prefix matches its sibling board-import.ts and states the subject
without leaning entirely on the folder name.
